requirements.txt
requests
pytest
python-dotenv
anthropic
google-genai
tenacity